+++ java-client/src/test/java/org/opensearch/client/opensearch/model/BehaviorsTest.java
@@ -42,9 +42,9 @@ import org.opensearch.client.opensearch._types.SortOptions;
 import org.opensearch.client.opensearch._types.SortOptionsBuilders;
 import org.opensearch.client.opensearch._types.SortOrder;
 import org.opensearch.client.opensearch._types.query_dsl.Query;
-import org.opensearch.client.opensearch._types.query_dsl.ShapeQuery;
 import org.opensearch.client.opensearch._types.query_dsl.TermQuery;
 import org.opensearch.client.opensearch._types.query_dsl.WrapperQuery;
+import org.opensearch.client.opensearch._types.query_dsl.XyShapeQuery;
 import org.opensearch.client.util.MapBuilder;
 
 public class BehaviorsTest extends ModelTestCase {
@@ -62,9 +62,9 @@ public class BehaviorsTest extends ModelTestCase {
 
     @Test
     public void testAdditionalPropertyOnClass() {
-        ShapeQuery q = new ShapeQuery.Builder().queryName("query-name")
+        XyShapeQuery q = new XyShapeQuery.Builder().queryName("query-name")
             .field("field-name")
-            .shape(_0 -> _0.relation(GeoShapeRelation.Disjoint))
+            .xyShape(_0 -> _0.relation(GeoShapeRelation.Disjoint))
             .ignoreUnmapped(true)
             .build();
 
@@ -72,7 +72,7 @@ public class BehaviorsTest extends ModelTestCase {
 
         assertEquals("query-name", q.queryName());
         assertTrue(q.ignoreUnmapped());
-        assertEquals(GeoShapeRelation.Disjoint, q.shape().relation());
+        assertEquals(GeoShapeRelation.Disjoint, q.xyShape().relation());
         System.out.println(toJson(q));
     }